feat(main): Add main page
The app haven't a main page to put other page into it.
This was a problem with navigation : we had to create
a special function with angular mecanism to set the
active page on the navigation. We cound't exploit all
the features provides by Ionic.
Now, this main page is called after the login page
and we put all others pages into it.
And we have to add an app controller in ParametersPage
for the logout system to return to the LoginPage if
user want to disconnect.